requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gspread==5.10.0
oauth2client==4.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7